[AF-424/#improve-performance-of-tests] -- updates

+++ tests/TestCase.php
@@ -3,11 +3,16 @@ namespace Tests;
 
 use Illuminate\Contracts\Console\Kernel;
 use Illuminate\Foundation\Testing\TestCase as BaseTestCase;
+//use Illuminate\Foundation\Testing\RefreshDatabase;
+use Symfony\Component\Console\Output\ConsoleOutput;
+
 use Tests\Asserts\CustomAsserts;
+use Database\Seeders\TestDatabaseSeeder;
 
 abstract class TestCase extends BaseTestCase
 {
     use CustomAsserts;
+    //use RefreshDatabase;
     //use CreatesApplication;
 
     public function createApplication()
@@ -21,13 +26,14 @@ abstract class TestCase extends BaseTestCase
 
     protected function setUp() : void
     {
+        parent::setUp();
+
         if ( !self::$isSetup ) {
+            //$this->output->writeln('TestCase -- calling setupDatabase...');
+            dump('TestCase -- calling setupDatabase...');
             $this->setupDatabase();
             self::$isSetup = true;
          }
-
-        parent::setUp();
-
     }
 
     public function ajaxJSON($method, $uri, array $data=[]) {
@@ -42,8 +48,22 @@ abstract class TestCase extends BaseTestCase
          ])->json($method,$uri,$data);
           */
     }
-    public function setupDatabase() {
-        exec('rm '.__DIR__.'/../database/tmp4test.sqlite');
-        exec('cp '.__DIR__.'/../database/template.sqlite '.__DIR__.'/../database/tmp4test.sqlite');
+
+    protected function setupDatabase() {
+        if (1) { 
+            //$this->output->writeln('setupDatabase -- copy template sqlite to test file...');
+            dump('setupDatabase -- copy template sqlite to test file...');
+            // use File:copy(?)
+            //  ~ see : https://laracasts.com/discuss/channels/laravel/unittest-with-pre-seeded-database-and-persistent-content
+
+            exec('rm '.__DIR__.'/../database/tmp4test.sqlite');
+            exec('cp '.__DIR__.'/../database/template.sqlite '.__DIR__.'/../database/tmp4test.sqlite');
+        } else {
+            //$this->output->writeln('setupDatabase -- seed memory database...');
+            dump('setupDatabase -- migrate memory database...');
+            \Artisan::call('migrate');
+            dump('setupDatabase -- seed memory database...');
+            $this->seed(TestDatabaseSeeder::class);
+        }
     }
 }
